maint: extracting ChatRequest.java
