+++ src/main/java/com/skool/zerotoprincipal/ThinkLikeAProgrammer/retrofit/Message.java
@@ -1,7 +1,6 @@
 package com.skool.zerotoprincipal.ThinkLikeAProgrammer.retrofit;
 
 import com.google.gson.annotations.SerializedName;
-import java.util.List;
 
 class Message {
     @SerializedName("role")
@@ -42,80 +41,3 @@ class Message {
     }
 }
 
-class ChatRequest {
-    @SerializedName("model")
-    private String model;
-
-    @SerializedName("messages")
-    private List<Message> messages;
-
-    @SerializedName("temperature")
-    private double temperature;
-
-    @SerializedName("max_tokens")
-    private int maxTokens;
-
-    @SerializedName("stream")
-    private boolean stream;
-
-    public ChatRequest(String model, List<Message> messages, double temperature, int maxTokens, boolean stream) {
-        this.model = model;
-        this.messages = messages;
-        this.temperature = temperature;
-        this.maxTokens = maxTokens;
-        this.stream = stream;
-    }
-
-    // Getters and setters
-
-    public String getModel() {
-        return model;
-    }
-
-    public void setModel(String model) {
-        this.model = model;
-    }
-
-    public List<Message> getMessages() {
-        return messages;
-    }
-
-    public void setMessages(List<Message> messages) {
-        this.messages = messages;
-    }
-
-    public double getTemperature() {
-        return temperature;
-    }
-
-    public void setTemperature(double temperature) {
-        this.temperature = temperature;
-    }
-
-    public int getMaxTokens() {
-        return maxTokens;
-    }
-
-    public void setMaxTokens(int maxTokens) {
-        this.maxTokens = maxTokens;
-    }
-
-    public boolean isStream() {
-        return stream;
-    }
-
-    public void setStream(boolean stream) {
-        this.stream = stream;
-    }
-
-    @Override
-    public String toString() {
-        return "ChatRequest{" +
-                "model='" + model + '\'' +
-                ", messages=" + messages +
-                ", temperature=" + temperature +
-                ", maxTokens=" + maxTokens +
-                ", stream=" + stream +
-                '}';
-    }
-}
